processing/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiolimiter==1.2.1
aioredis==2.0.1
aiosignal==1.3.1
aiostream==0.5.2
altair==5.2.0
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.2.0
anytree==2.12.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.2
attrs==23.2.0
bce-python-sdk==0.9.25
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.28.21
botocore==1.31.21
bpemb==0.3.4
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.2
camel-converter==4.0.1
certifi==2023.11.17
cffi==1.15.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmudict==1.0.32
comm==0.2.2
conllu==4.5.3
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dashscope==1.14.1
dataclasses-json==0.6.3
datasets==2.14.3
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distro==1.9.0
Django==5.1.5
dnspython==2.4.2
et_xmlfile==2.0.0
exceptiongroup==1.1.2
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.109.0
fastjsonschema==2.21.1
filelock==3.12.2
fire==0.4.0
firebase-admin==6.2.0
flair==0.12.2
fonttools==4.42.0
frozenlist==1.4.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==4.3.1
gitdb==4.0.11
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gotrue==2.11.3
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.56.0
grpcio-tools==1.62.3
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.16.4
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.6
imap-tools==1.5.0
importlib-metadata==7.0.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
Janome==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lancedb==0.4.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-experimental==0.0.49
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.50
lazy-object-proxy==1.10.0
libcst==1.0.1
loguru==0.6.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
meilisearch==0.21.0
metagpt==0.8.1
modal==0.65.43
more-itertools==10.1.0
motor==3.6.1
mpld3==0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nextcord==2.4.2
numpy==1.24.3
openai==1.6.1
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openpyxl==3.1.5
orjson==3.10.1
outcome==1.2.0
overrides==7.7.0
packaging==23.2
pandas==2.1.1
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
playwright==1.49.1
pluggy==1.5.0
portalocker==2.10.1
postgrest==0.19.3
pptree==3.1
prance==23.6.21.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.22.3
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.7
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyee==12.0.0
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.9.2
pyparsing==3.0.9
pyphen==0.17.2
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytorch_revgrad==0.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.7.0
qianfan==0.3.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
realtime==2.3.0
referencing==0.30.2
regex==2023.12.25
reportlab==4.2.5
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.2
scipy==1.11.1
segtok==1.5.11
selenium==4.11.2
semantic-kernel==0.4.3.dev0
semver==3.0.4
sentencepiece==0.1.99
sentry-sdk==1.39.1
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlitedict==2.1.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.35.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.12
synchronicity==0.9.3
ta==0.10.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.5.0
textstat==0.7.5
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformer-smaller-training-vocab==0.2.4
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.27.0
validators==0.22.0
watchfiles==0.24.0
wcwidth==0.2.6
webdriver-manager==4.0.2
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
Wikipedia-API==0.6.0
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.18.3
zhipuai==2.0.1
zipp==3.17.0
